# 06_IMPLEMENTATION/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Marpex CRM API
  version: 0.1.0
  description: |
    Základná OpenAPI špecifikácia pre MVP verziu Marpex CRM.
servers:
  - url: http://localhost:3005
    description: Local API
...
paths:
...
  /api/visits:
    get:
      tags: [Visits]
      summary: List visits
      parameters:
        - in: query
          name: customerId
          schema:
            type: string
            format: uuid
        - in: query
          name: ownerId
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: late
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Visit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'
    post:
      tags: [Visits]
      summary: Create visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitInput'
      responses:
        '201':
          description: Created visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          $ref: '#/components/responses/ApiError'
  /api/opportunities:
    get:
      tags: [Opportunities]
      summary: List opportunities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Opportunity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'
    post:
      tags: [Opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityInput'
      responses:
        '201':
          description: Created opportunity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '400':
          $ref: '#/components/responses/ApiError'
  /api/opportunities/{id}/stage:
    patch:
      tags: [Opportunities]
      summary: Move opportunity stage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage]
              properties:
                stage:
                  type: string
                  enum: [identified_need, qualified, technical_solution, quote_delivered, negotiation, verbal_confirmed, won, lost]
                technicalSpec:
                  type: string
                competition:
                  type: string
                followUpDate:
                  type: string
                  format: date
                closeResult:
                  type: string
                closeTimestamp:
                  type: string
                  format: date-time
                lostReason:
                  type: string
      responses:
        '200':
          description: Stage moved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  newStage:
                    type: string
                    example: qualified
        '400':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - in: query
          name: opportunityId
          schema:
            type: string
            format: uuid
        - in: query
          name: customerId
          schema:
            type: string
            format: uuid
        - in: query
          name: done
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, dueDate]
              properties:
                title:
                  type: string
                dueDate:
                  type: string
                  format: date
                opportunityId:
                  type: string
                  format: uuid
                customerId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
components:
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
      description: Voliteľné číslo stránky.
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Voliteľný page size.
  responses:
    ApiError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validation:
              value:
                error: Validation error
                code: VALIDATION_ERROR
                details:
                  field:
                    - Povinné pole
  schemas:
    ApiError:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Visit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        visitGoal:
          type: string
        result:
          type: string
        potentialEur:
          type: string
        nextStep:
          type: string
        nextStepDeadline:
          type: string
          format: date
    VisitInput:
      type: object
      required:
        - date
        - customerId
        - contactId
        - visitGoal
        - result
        - customerNeed
        - opportunityCreated
        - potentialEur
        - competition
        - nextStep
        - nextStepDeadline
      properties:
        date:
          type: string
          format: date
        customerId:
          type: string
          format: uuid
        contactId:
          type: string
          format: uuid
        visitGoal:
          type: string
        result:
          type: string
        customerNeed:
          type: string
        opportunityCreated:
          type: boolean
        opportunityType:
          type: string
          enum: [project, service, cross_sell]
        potentialEur:
          type: number
        competition:
          type: string
        nextStep:
          type: string
        nextStepDeadline:
          type: string
          format: date
    Opportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        customerId:
          type: string
          format: uuid
        value:
          type: string
        stage:
          type: string
        nextStepSummary:
          type: string
        nextStepDeadline:
          type: string
          format: date
    OpportunityInput:
      type: object
      required: [title, customerId, value, stage, nextStepSummary, nextStepDeadline]
      properties:
        title:
          type: string
          example: Retrofit linky ABC
        customerId:
          type: string
          format: uuid
        value:
          type: number
          example: 45000
        stage:
          type: string
          enum: [identified_need, qualified, technical_solution, quote_delivered, negotiation, verbal_confirmed]
        nextStepSummary:
          type: string
          example: Poslať technický návrh
        nextStepDeadline:
          type: string
          format: date
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
          nullable: true